requirements.txt
click==8.0.3
Flask==2.0.2
Flask-Reuploaded==1.2.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.5
opencv-python==4.5.4.60
Pillow==8.4.0
pytesseract==0.3.8
Werkzeug==2.0.2
